daemon: Implement vchan_init function and related helpers
This commit implements the vchan_init() call that ties together
the daemon parent process (main), the ivshmem loop, the libvirt loop,
and a newly introduced localhandler to allow an external process
to request the creation of a vchan between two domains.

The following IPC calls were implemented:
    * LIBVIRT_IPC_CMD_ATTACH_IVSHMEM
        - Request a new ivshmem device to be attached to the given
        domain(s).

    * IVSHMEM_IPC_CMD_REGISTER_CONN
        - Register a new upcomming ivshmem device connection
        from the given pid(s) that will be backed by the
        given memfd.

Tied together in the implementation of vchan_init(), the daemon now
provides the baseline functionality required to create arbitrary vchans.
The ability to destroy vchans is not yet implemented.

The localhandler loop allows external processes to interface with
the daemon via the stripped-down IPC protocol defined in
libkvmchan-priv.h. This is what libkvmchan.so will use to provide
the Xen vchan API to normal programs. Currently the
KVMCHAND_CMD_SERVERINIT is implemented which will simply forward
all arguments to the daemon's internal MAIN_IPC_CMD_VCHAN_INIT.