Implement library wrappers for open/read/write operations
The libkvmchan.so library now implements wrappers for all
of kvmchand's supported client operations, which includes
most of the libvchan API. The ringbuf implementation was
amended to support both streaming and packet-based modes
of communication, as required by libvchan.

A simple demo, test_library.c, was introduced to demonstrate
all functionality currently exposed by libkvmchan.so.

Notably missing is the ability to close vchans. This will
require changes to kvmchand to allow detaching ivshmem devices
and deallocating their resources.