library: Add _data_ready and _buffer_space, and start _wait
Implement _data_ready and _buffer_space as simple wrappers for
internal ringbuf functions.

In addition to those, this commit adds initial support for vchan
state management. When a client vchan is connected, it will send
a dummy byte to indicate to the server that a connection has been
established. The server will check for this bit and set an internal
flag to indicate that the client has connected.

Before the client has connected, all server vchan operations will
block. Once a way to determine vchan close has been implemented,
libvchan_wait can be implemented.