Moved systemd sd_notify() from vfio.c to localhandler.c
A systemd unit file that has the 'Type' set to 'notify' enables the ability for
systemd to notify any unit that 'Requires' it when a call is made to
sd_nofify().  The unit file that requires to be notified will not start until
systemd recieves a 'READY=1' message which allows services that rely on another
service to be fully initialized to wait preventing race conditions.

The kvmchand service unit file service 'Type' was changed from 'simple' to
'notify' and dropins were added for qubes-db.service and qubes-qrexec.service
to `Require` the kvmchand.service unit to force them to wait until kvmchan
initialization is complete since the services were starting too early which
caused fatal communication errors that could not recover.