Reuse vchans between domains
This commit reworks vchan management in the daemon to keep vchans
between domains even after they're closed. These vchans are then
reused for future connections on a first-fit basis, i.e. the first
reusable vchan with enough space is used.

When a suitable vchan isn't available for reuse, a new vchan is created.