Implement blocking in libkvmchan_client_init
Make libkvmchan_client_init block and retry indefinitely as long as the
specified server dom is online. This more closely matches the Xen vchan
behavior. Closes #20.

To facilitate this behavior, vchan_conn now returns a status code
indicating why the connection failed - due to the domain being offline
or the vchan being unavailable. This status code is propagated back to
local kvmchand clients and library.c's libkvmchan_client_init uses it to
determine whether it should retry or not.