daemon: Fix race between localhandler and ivshmem client requests
In the complicated web of subprocesses and IPC that is kvmchand,
client requests that propagate throughout the entire daemon can be
initiated from two places:
    * localhandler.c for clients on the local domain using the unix
      socket API.
and
    * (host only) ivshmem.c for clients on guest domains that have
      kvmchand-guest instances running and communicate.

Both of these request entrypoints end up calling the same internal
IPC endpoints, none of which were designed with reentrancy in mind.
This creates an obvious issue in the case where a request originates
from one entrypoint and before it is finished, another request originates
from the other entrypoint. This has the potential to cause Bad Things to
happen (namely, daemon-wide deadlocks).

This issue was not discovered during initial development and testing which
used simple, manually-initiated test programs. When trying with a real Qubes
environment though, the vast number of vchans and daemons that all start up
when a VM is launch resulted in the race being hit quite frequently.

There would be two routes to fix this issue: Modifying all IPC endpoints
to support reentrancy or adding a global request lock that ensures only
one entrypoint is servicing a request at a time. As the former would require
a massive overhaul of the daemon, this commit implements the latter.

Locking is implemented with an eventfd in EFD_SEMAPHORE mode that is created
by localhandler and passed to ivshmem over IPC during early init. Both client
request event loops were then modified to acquire a lock on the eventfd before
servicing a given request and to drop it afterwards. This change was minimally
invasive and seems to have completely resolved all deadlocks during testing.