daemon: Fix localhandler crash when in guest mode
Commit 5e81fa85cb12 ("daemon: Fix race between localhandler and
ivshmem client requests") introduced a bug where localhandler.c would
cause an IPC crash when run in guest mode. This is due to code it
added that would unconditionally send a message to ivshmem.c to set
up the client request lock; this works fine on dom0 but in guest mode
there is no ivshmem.c process running so this results in an IPC bug
assertion.

Fix the issue by guarding the client request lock setup by a check for
`is_dom0` so that the wrong code paths aren't executed in guest mode.