daemon: Expand types of vchans that can be reused
kvmchand has the ability to reuse vchans that have been closed,
but the previous logic would only do this for vchans where the new
server equals the old server and the new client equals the old client.
This excludes vchans where the new server is the old client and vice
versa, even though there is no technical reason for doing os.

This commit expands the reuse logic so that directionality is not
taken into account when selecting vchans for reuse. This should result
in more efficient resource usage.