library: Work around race in libkvmchan_client_init
Work around a race in libkvmchan_client_init where the magic value
of a newly created vchan is checked before the server has had the
time to initialize it properly. This resulted in false-negative returns
and can break applications that only try client_init once and rely on
success after blocking.