daemon: Refactor management of vchan connections
Previously, all vchan connections between domains were recorded in
a single flat vector. This commit refactors the data structure so
vchan connections are instead stored in a linked list owned by the
connection's server domain, and each currently active domain is
stored in a global linked list.

This new hierarchical layout is a more logical way to represent and
retrieve connection information and will make it easier to track
shared pages when the shared page API is implemented.

A generic doubly linked list implementation was added to util{.c,.h}
for this.