Initial implementation of shared memory API
This commit introduces major changes across the daemon and library to
lay the groundwork for supporting a new shared memory API. The goal is
to provide the functionality required for implementing the Qubes GUI
agent and daemon which currently use Xen-specific APIs for sharing
pages.

Currently, only the shmem_create function is implemented, which creates
a backing ivshmem shared memory device between two domains, allocates
the requested number of pages in it, then returns information on the
mapping to the caller. The information includes a region id, an
ivposition (if applicable), an mmap-able file descriptor, and an offset
within the file descriptor where the region begins.

Design
------
To reduce resource usage, a new approach for ivshmem device allocation
was implemented for this API. Specifically, shmem allocations are not
given unique individual ivshmem devices, but instead are allocated
portions of shared ivshmem memory pools. This way only a few memory
pools need to be allocated to support a number of shmem mappings, and
the losses from needing to round up to powers of two for PCI BAR sizes
are not relevant.

The main tradeoff here is that processes can snoop on other process'
shared memory allocations by ignoring the kvmchand-provided offset when
they map their file descriptors. This seems like a reasonable tradeoff
since processes are limited to snooping on memory shared with their
domain, but if this is deemed to an issue, access control to memory
pools can be added.

Overview of changes
-------------------

daemon/connections:
    * Reorganize connection database yet again - both connection structs
      and domain structs are stored in top-level linked lists.
      Connection structs contain pointers to relevant domain structs.

    * Move boilerplate for creating new ivshmem devices out of
      vchan_init into a generic helper function.

    * Add `shmem_create` function, analogue of `vchan_init` for the new
      shmem API.

daemon/localhandler:
    * Rework handling of client handle management to make it easier to
      add future handle types.

    * Lay groundwork for keeping track of created shmem handles.
      Cleaning them up isn't implemented yet though.

    * Support for servicing KVMCHAND_CMD_SHMEM_CREATE for dom0 clients.

daemon/page_allocator:
    * Add new generic page allocator API, used for dividing up memory
      pools into individual allocations. This is pretty barebones and
      probably has some rough edges that need to be fixed.

daemon/ipc:
    * Add MAIN_IPC_CMD_SHMEM_CREATE endpoint, modify behavior of
      *_GET_CONN_FDS endpoints to allow only retrieving the memfd.

libkvmchan-priv.h:
    * Add KVMCHAND_CMD_SHMEM_CREATE endpoint.
    * Cleanup documentation of other endpoints.

library.c:
    * Add wrapper for shmem_create.

And a few other miscellaneous changes.