Implement shmem_close, support calling shmem_* from guests
This commit introduces support for the shmem_close operation for
server-side closing of previously created shmem regions. In addition, it
adds support for calling shmem_ functions from guest domains.

A variety of cleanups and bugfixes to both the llist implementation,
page allocator, and connection database were added to facilitate this.