daemon: Simplify MAIN_IPC_CMD_SHMEM_CREATE to return memfd
This removes the need for an additional IPC call to
IVSHMEM_IPC_CMD_GET_CONN_FDS when creating shmem regions on dom0,
and simplifies the logic as a result.

It also removes the need to return the qemu pid for the peer which
means only 2 IPC `ret` files are necessary again, so remove ret3/4.