daemon/page_allocator: Fix chunk free w/ no coalesce
Previously if a chunk was freed but couldn't be coalesced forward or
backward it would be removed from the chunk list. This behavior is
incorrect and creates a hole in the region's memory map - on free if
coalescing is not possible, the chunk should simply be marked as free
and kept in the list.