library: Implement wrappers for full shmem API
This includes fixes for mapping shmem regions from guests when offset>0.
This also means we have to do away with the mmap parameter API and
instead need to map all memory ourselves, but this was deemed acceptable
by Qubes upstream for the time being.

Also include some more documentation in the form of comments for shmem
API functions in the header.